Revert "18dds 17"
