+++ src/main/java/at/htl/persistence/entity/Room.java
@@ -1,10 +1,5 @@
 package at.htl.persistence.entity;
 
-import at.htl.persistence.dao.RoomDao;
-import at.htl.persistence.dao.TeacherDao;
-import at.htl.rest.dto.LessonDto;
-import at.htl.rest.dto.RoomDto;
-import at.htl.rest.util.Weekday;
 import javax.persistence.*;
 import javax.xml.bind.annotation.XmlRootElement;
 import java.util.List;
@@ -32,19 +27,6 @@ public class Room {
         this.name = name;
     }
     //endregion
-
-    public RoomDto toDto(){
-        RoomDto dto = new RoomDto();
-        dto.setId(id);
-        if(name != null)
-            dto.setName(name);
-        return dto;
-    }
-
-    public void update(RoomDto roomDto) {
-        if(roomDto.getName() != null)
-            this.name = roomDto.getName();
-    }
 
     //region Getters and Setters
     public Integer getId() {
