+++ src/main/java/at/htl/persistence/entity/Teacher.java
@@ -1,8 +1,7 @@
 package at.htl.persistence.entity;
 
-import at.htl.rest.dto.TeacherDto;
-
 import javax.persistence.*;
+import javax.xml.bind.annotation.XmlRootElement;
 import java.util.List;
 
 @Entity
@@ -15,8 +14,7 @@ public class Teacher {
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "Teacher_Seq")
     @SequenceGenerator(name = "Teacher_Seq", sequenceName = "Teacher_Seq")
     private Integer id;
-    private String lastName;
-    private Boolean isMale;
+    private String name;
     @OneToMany
     private List<Lesson> lessons;
 
@@ -24,26 +22,12 @@ public class Teacher {
     public Teacher() {
     }
 
-    public Teacher(Integer id, String lastName) {
+    public Teacher(Integer id, String name) {
         this.id = id;
-        this.lastName = lastName;
+        this.name = name;
     }
     //endregion
 
-    public TeacherDto toDto(){
-        TeacherDto dto = new TeacherDto();
-        dto.setId(id);
-        if(lastName != null)
-            dto.setLastName(lastName);
-        if(isMale != null)
-            dto.setIsMale(isMale);
-        return dto;
-    }
-
-    public void update(TeacherDto teacherDto) {
-        this.lastName = teacherDto.getLastName();
-        this.isMale = teacherDto.getIsMale();
-    }
     //region Getters and Setters
     public Integer getId() {
         return id;
@@ -53,24 +37,12 @@ public class Teacher {
         this.id = id;
     }
 
-    public String getLastName() {
-        return lastName;
+    public String getName() {
+        return name;
     }
 
-    public void setLastName(String name) {
-        this.lastName = name;
-    }
-
-    public Boolean getIsMale() {
-        return isMale;
-    }
-
-    public void setIsMale(Boolean is_male) {
-        this.isMale = is_male;
-    }
-
-    public void setLessons(List<Lesson> lessons) {
-        this.lessons = lessons;
+    public void setName(String name) {
+        this.name = name;
     }
 
     public List<Lesson> getLessons(){ return this.lessons;}
