--- src/main/java/at/htl/rest/dto/UnitDto.java
@@ -1,44 +1,0 @@
-package at.htl.rest.dto;
-
-import javax.xml.bind.annotation.XmlRootElement;
-import java.sql.Timestamp;
-
-@XmlRootElement
-public class UnitDto {
-    private Integer id;
-    private Timestamp startTime;
-    private Timestamp endTime;
-
-    public UnitDto() {
-    }
-
-    public UnitDto(Integer id, Timestamp startTime, Timestamp endTime) {
-        this.id = id;
-        this.startTime = startTime;
-        this.endTime = endTime;
-    }
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public Timestamp getStartTime() {
-        return startTime;
-    }
-
-    public void setStartTime(Timestamp startTime) {
-        this.startTime = startTime;
-    }
-
-    public Timestamp getEndTime() {
-        return endTime;
-    }
-
-    public void setEndTime(Timestamp endTime) {
-        this.endTime = endTime;
-    }
-}
