+++ src/main/java/at/htl/rest/endpoint/LessonEndpoint.java
@@ -1,8 +1,10 @@
 package at.htl.rest.endpoint;
 
-import at.htl.persistence.dao.*;
+import at.htl.persistence.dao.Dao;
+import at.htl.persistence.dao.LessonDao;
+import at.htl.persistence.dao.RoomDao;
+import at.htl.persistence.dao.TeacherDao;
 import at.htl.persistence.entity.Lesson;
-import at.htl.persistence.entity.Subject;
 import at.htl.rest.dto.LessonDto;
 
 import javax.inject.Inject;
@@ -21,8 +23,6 @@ public class LessonEndpoint extends EntityEndpoint<Lesson, LessonDto>{
     RoomDao roomDao;
     @Inject
     TeacherDao teacherDao;
-    @Inject
-    SubjectDao subjectDao;
 
     @Override
     protected Dao<Lesson> getEntityDao() {
@@ -41,7 +41,7 @@ public class LessonEndpoint extends EntityEndpoint<Lesson, LessonDto>{
 
     @Override
     protected Lesson updateEntityWithDto(Lesson lesson, LessonDto lessonDto) {
-        lesson.update(lessonDto, roomDao, teacherDao, subjectDao);
+        lesson.update(lessonDto, roomDao, teacherDao);
         return lesson;
     }
 }
